+++ src/QNEthernet.cpp
@@ -346,13 +346,6 @@ void EthernetClass::setDNSServerIP(const IPAddress &dnsServerIP) {
   DNSClient::setServer(0, dnsServerIP);
 }
 
-bool EthernetClass::sendRaw(const uint8_t *frame, size_t len) {
-  if (netif_ == nullptr) {
-    return false;
-  }
-  return enet_output_frame(frame, len);
-}
-
 bool EthernetClass::joinGroup(const IPAddress &ip) {
   if (netif_ == nullptr) {
     return false;