+++ src/QNEthernet.h
@@ -156,19 +156,6 @@ class EthernetClass final {
   [[deprecated]] void setRetransmissionCount(uint8_t number) {}
   [[deprecated]] void setRetransmissionTimeout(uint16_t milliseconds) {}
 
-  // Sends a raw Ethernet frame. This returns whether the send was successful.
-  //
-  // The FCS (Frame Check Sequence, the CRC value) should not be included in the
-  // frame data.
-  //
-  // This will return false if:
-  // 1. Ethernet was not started,
-  // 2. The frame is NULL, or
-  // 3. The length is not in the range 60-1518.
-  //
-  // See also: EthernetFrame API
-  bool sendRaw(const uint8_t *frame, size_t len);
-
   // Joins a multicast group. This returns whether the call was successful.
   //
   // The lwIP stack keeps track of a group "use count", so this function can be